JSON: holder has empty() function, moved value_utils to value.h
