+++ src/libstrtb/config/system.h
@@ -3,7 +3,6 @@
 
 #include "id_type.h"
 #include "../json/value.h"
-#include "../json/value_utils.h"
 #include "../logging/logging.h"
 
 #include <string>
