+++ src/libstrtb/json/value_array.h
@@ -2,7 +2,7 @@
 #define STRTB_JSON_VALUE_ARRAY_H
 
 #include "value.h"
-#include "value_utils.h"
+
 
 #include <vector>
