+++ src/libstrtb/json/value_object.h
@@ -2,7 +2,7 @@
 #define STRTB_JSON_VALUE_OBJECT_H
 
 #include "value.h"
-#include "value_utils.h"
+
 
 #include <map>
 #include <string>
